clean up check registration, error implementation
