+++ pandera/backends/__init__.py
@@ -1,7 +1,7 @@
 """Pandera backends."""
 
 # ensure that base builtin checks and hypothesis are registered
-import pandera.core.base.builtin_checks
-import pandera.core.base.builtin_hypotheses
+import pandera.backends.stubs.builtin_checks
+import pandera.backends.stubs.builtin_hypotheses
 
 import pandera.backends.pandas
