+++ pandera/extensions.py
@@ -2,8 +2,8 @@
 
 # pylint: disable=unused-import
 from pandera.core.extensions import (
-    register_check,
-    register_hypothesis,
+    register_builtin_check,
+    register_builtin_hypothesis,
     generate_check_signature,
     generate_check_annotations,
     modify_check_fn_doc,
